Fix `collect_from_directory_symlink` and remove `cd` variants
Now `collect_from_directory_symlink` does not give error when
encountering a symlink, the problem was, I was doing a `path.exists()`
check which actually follows symlinks, so it was telling my symlink
files didn't exist when the links were broken.

I removed this check, and now checks are done only with functions that
correctly treat symlinks (don't follow them).